pyproject.toml
[project]
name = "thetagang-notifications"
version = "2.2.0"
description = "Notifications for the Theta Gang discord"
readme = "README.md"
authors = [{ name = "Major Hayden", email = "major@mhtx.net" }]
requires-python = ">=3.13,<4.0"
dependencies = [
    "discord-webhook>=1.4.1",
    "httpx>=0.28.1",
    "inflect>=7.5.0",
    "python-dateutil>=2.9.0.post0",
    "ruyaml>=0.91.0",
    "redis>=6.4.0",
    "schedule>=1.2.2",
    "pydantic-settings>=2.8.2",
]

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "fakeredis>=2.31.1",
    "freezegun>=1.5.5",
    "ipython>=9.5.0",
    "pyright>=1.1.405",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.0",
    "pytest-recording>=0.13.4",
    "ruff>=0.12.12",
    "types-python-dateutil>=2.9.0.20250822",
    "types-PyYAML>=6.0.12.20250822",
    "types-redis>=4.6.0.20241004",
    "vcrpy>=7.0.0",
]


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--cov=thetagang_notifications",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-branch",
    "--color=yes",
]
# Having problems with tests? Uncomment the following line to see more output.
# log_cli = false
# log_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = ["tests"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
fix = true

[tool.ruff.format]
preview = true
